refactor: 이미지 업로드 플랫폼 변경 및 OpenFeign 설정 수정 #18 - 이미지 업로드 플랫폼 s3 -> GCS로 변경 - s3 관련 필드명 수정 s3Url -> imageUrl - OpenFeignConfig 위치를 common 디렉토리로 변경
